Refactor evaluation elsewhere
