+++ packages/tables/src/Filters/BaseFilter.php
@@ -10,13 +10,14 @@ class BaseFilter extends Component
     use Concerns\BelongsToTable;
     use Concerns\CanBeHidden;
     use Concerns\CanSpanColumns;
-    use Concerns\EvaluatesClosures;
     use Concerns\HasDefaultState;
     use Concerns\HasFormSchema;
     use Concerns\HasLabel;
     use Concerns\HasName;
     use Concerns\InteractsWithTableQuery;
     use Conditionable;
+
+    protected string $evaluationIdentifier = 'filter';
 
     final public function __construct(string $name)
     {
@@ -29,5 +30,12 @@ class BaseFilter extends Component
         $static->setUp();
 
         return $static;
+    }
+
+    protected function getDefaultEvaluationParameters(): array
+    {
+        return array_merge(parent::getDefaultEvaluationParameters(), [
+            'livewire' => $this->getLivewire(),
+        ]);
     }
 }
