--- packages/tables/src/Filters/Concerns/EvaluatesClosures.php
@@ -1,28 +1,0 @@
-<?php
-
-namespace Filament\Tables\Filters\Concerns;
-
-use Closure;
-
-trait EvaluatesClosures
-{
-    public function evaluate($value, array $parameters = [])
-    {
-        if ($value instanceof Closure) {
-            return app()->call(
-                $value,
-                array_merge($this->getDefaultEvaluationParameters(), $parameters),
-            );
-        }
-
-        return $value;
-    }
-
-    protected function getDefaultEvaluationParameters(): array
-    {
-        return [
-            'filter' => $this,
-            'livewire' => $this->getLivewire(),
-        ];
-    }
-}
